Move frontend-specific calltypes to /rpc
