+++ src/node/calltypes.h
@@ -61,21 +61,4 @@ namespace ccf
       std::vector<uint8_t> quote;
     };
   };
-
-  struct GetCommit
-  {
-    struct Out
-    {
-      uint64_t term;
-      int64_t commit;
-    };
-  };
-
-  struct GetTxHist
-  {
-    struct Out
-    {
-      nlohmann::json tx_hist;
-    };
-  };
 }
